fix: android installation and path settings
The dotfiles repository contains helpers for handling path appending. However, on fish, paths can't be appended if they do not exist yet. As a workaround, instead of globally appending to path once we append to path as soon as we can. The main culpit was the android tool installation and its path settings.